+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -22,6 +22,8 @@
   ([#41950](https://github.com/Azure/azure-sdk-for-python/pull/41950))
 - Customer Facing Statsbeat: Added logic for retry item count
   ([#41971](https://github.com/Azure/azure-sdk-for-python/pull/41971))
+- Customer Facing Statsbeat: Updated base and storage to include the metrics for customer statsbeat
+  ([#42123](https://github.com/Azure/azure-sdk-for-python/pull/42123))
 
 
 - Support AI Foundry by Handling GEN_AI_SYSTEM Attributes with [Spec](https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/blob/main/ApplicationInsights/genai_semconv_mapping.md) ([#41705](https://github.com/Azure/azure-sdk-for-python/pull/41705))